[#62809724] [#63120498] [#62820502]
[#62809724] [#63120498]- Now the TypeError will not be raised and it is
handled in wos.py and the error raised because of MAP won't occur
again.
Though the fix is ready for these 2 stories, working on the legible way
of throwing error when the input data is not in the correct format.

[#62820502] The remaining pylint errors are in most of the unused
scripts like bib.py and so on. Except those, most of the genuine cases
were cleared.